pyproject.toml
[tool.poetry]
name = "mini-expense-tracker"
version = "0.1.0"
description = ""
authors = ["Peter Laudel <38426989+PeterLaudel@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
alembic = "^1.14.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
factory-boy = "^3.3.1"
mypy = "^1.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
